Refactor with separate objects
Try to work around symbol/export trickery
