+++ ci/rebuild-libwit-bindgen-cabi.sh
@@ -42,10 +42,11 @@ set -ex
 version=$(./ci/print-current-version.sh | sed 's/\./_/g')
 
 realloc=cabi_realloc_wit_bindgen_$version
-wasip3_task_set=wasip3_task_set_wit_bindgen_$version
-wasip3_task_get=wasip3_task_get_wit_bindgen_$version
 
-cat >./crates/guest-rust/rt/src/wit_bindgen_cabi.rs <<-EOF
+rm -f crates/guest-rust/rt/src/wit_bindgen_*.{rs,o,c}
+rm -f crates/guest-rust/rt/src/libwit_bindgen_cabi.a
+
+cat >./crates/guest-rust/rt/src/wit_bindgen_cabi_realloc.rs <<-EOF
 // This file is generated by $0
 
 #[unsafe(no_mangle)]
@@ -57,20 +58,9 @@ pub unsafe extern "C" fn $realloc(
 ) -> *mut u8 {
     crate::cabi_realloc(old_ptr, old_len, align, new_len)
 }
-
-static mut WASIP3_TASK: *mut u8 = core::ptr::null_mut();
-
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn $wasip3_task_set(ptr: *mut u8) -> *mut u8 {
-    unsafe {
-        let ret = WASIP3_TASK;
-        WASIP3_TASK = ptr;
-        ret
-    }
-}
 EOF
 
-cat >./crates/guest-rust/rt/src/wit_bindgen_cabi.c <<-EOF
+cat >./crates/guest-rust/rt/src/wit_bindgen_cabi_realloc.c <<-EOF
 // This file is generated by $0
 
 #include <stdint.h>
@@ -81,25 +71,37 @@ __attribute__((__weak__, __export_name__("cabi_realloc")))
 void *cabi_realloc(void *ptr, size_t old_size, size_t align, size_t new_size) {
   return $realloc(ptr, old_size, align, new_size);
 }
+EOF
 
-extern void *$wasip3_task_set(void *ptr);
+cat >./crates/guest-rust/rt/src/wit_bindgen_cabi_wasip3.c <<-EOF
+// This file is generated by $0
+
+#include <stdlib.h>
+
+static void *WASIP3_TASK = NULL;
 
 __attribute__((__weak__))
 void *wasip3_task_set(void *ptr) {
-  return $wasip3_task_set(ptr);
+  void *ret = WASIP3_TASK;
+  WASIP3_TASK = ptr;
+  return ret;
 }
 EOF
 
-rm -f crates/guest-rust/rt/src/wit_bindgen_cabi.o
-$WASI_SDK_PATH/bin/clang crates/guest-rust/rt/src/wit_bindgen_cabi.c \
-  -O -c -o crates/guest-rust/rt/src/wit_bindgen_cabi.o
+build() {
+  file=$1
+  $WASI_SDK_PATH/bin/clang crates/guest-rust/rt/src/$1.c \
+    -O -c -o crates/guest-rust/rt/src/$1.o
+  # Remove the `producers` section. This appears to differ whether the host for
+  # clang is either macOS or Linux. Not needed here anyway, so discard it to help
+  # either host produce the same object.
+  wasm-tools strip -d producers ./crates/guest-rust/rt/src/$1.o \
+    -o ./crates/guest-rust/rt/src/$1.o
+}
 
-# Remove the `producers` section. This appears to differ whether the host for
-# clang is either macOS or Linux. Not needed here anyway, so discard it to help
-# either host produce the same object.
-wasm-tools strip -d producers ./crates/guest-rust/rt/src/wit_bindgen_cabi.o \
-  -o ./crates/guest-rust/rt/src/wit_bindgen_cabi.o
+build wit_bindgen_cabi_realloc
+build wit_bindgen_cabi_wasip3
 
-rm -f crates/guest-rust/rt/src/libwit_bindgen_cabi.a
 $WASI_SDK_PATH/bin/llvm-ar crus crates/guest-rust/rt/src/libwit_bindgen_cabi.a \
-  crates/guest-rust/rt/src/wit_bindgen_cabi.o
+  crates/guest-rust/rt/src/wit_bindgen_cabi_realloc.o \
+  crates/guest-rust/rt/src/wit_bindgen_cabi_wasip3.o
