+++ crates/guest-rust/rt/src/wit_bindgen_cabi_realloc.c
@@ -1,20 +1,10 @@
 // This file is generated by ./ci/rebuild-libwit-bindgen-cabi.sh
 
 #include <stdint.h>
-#include <stdlib.h>
 
 extern void *cabi_realloc_wit_bindgen_0_41_0(void *ptr, size_t old_size, size_t align, size_t new_size);
 
 __attribute__((__weak__, __export_name__("cabi_realloc")))
 void *cabi_realloc(void *ptr, size_t old_size, size_t align, size_t new_size) {
   return cabi_realloc_wit_bindgen_0_41_0(ptr, old_size, align, new_size);
-}
-
-static void *WASIP3_TASK = NULL;
-
-__attribute__((__weak__))
-void *wasip3_task_set(void *ptr) {
-  void *ret = WASIP3_TASK;
-  WASIP3_TASK = ptr;
-  return ret;
 }
